+++ src/nodes/Hudu/optionLoaders/index.ts
@@ -2,3 +2,7 @@ export * from './users';
 export * from './companies';
 export * from './asset_layouts';
 export * from './lists';
+export * from './groups';
+export * from './networks';
+export * from './vlans';
+export * from './vlan_zones';